requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
iniconfig==2.0.0
mysqlclient==2.2.1
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
PyJWT==2.8.0
pytest==8.0.0
pytz==2023.4
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
